Added get_url_from_name
Function returns url from name.
Also updated the image naming function to prevent duplicate names.

NOTE: this does cause a bug where you can upload an image with a name, and the image url is uploaded to the database but not the image name, because of duplicate names.

This will be fixed in the next update, pushing this code because someone needs it.